ofed_1_4_scripts: Fix ppc64 building of rnfs-utils
rnfs-utils will not build on ppc64 without a 32bit version of
e2fsprogs-devel.  The 32bit version is needed due to logic previously
added to install.pl to force the compile to be 32bit.  Another way to
fix this issue would be to allow 64bit version of the rnfs-utils by
removing the logic which forces it to be 32bit.

With the added dependency, rnfs-utils compiles on RHEL5.2 on ppc64
without issue.

Signed-Off-By: Jon Mason <jon@opengridcomputing.com>