install scripts: rnfs-utils SLES10sp2 build break fix
Building rnfs-utils on SLES10sp2 breaks due to tcpd-devel issues with
32/64bit libraries.  tcpd-devel is only needed for TCP Wrapper
enablement in rnfs-utils.  This is unnecessary, as TCP Wrappers are only
needed on the server, and rnfs-utils is only need on the client.

With rnfs-utils-1.1.5-2.OFED, tcp-wrappers are disabled and the
build-req is no longer necessary.

Also, e2fsprogs-devel is needed by ppc in addition to ppc64.

This patch fixes bugs 1592 and 1594.

Signed-Off-By: Jon Mason <jon@opengridcomputing.com>