install.pl: rework logic determining whether to allow NFSRDMA
It is possible to install NFSRDMA on unsupported distros by installing a
newer kernel on an older distro.  Unfortunately, this causes issues with
trying to instal rnfs-utils and potentially other nasty side effects.
By reducing the logic to check for the supported distros and the
supported kernels only, then this should it prevent from happening.
Also, the logic is much cleaner this way.

This fixes bug 1825

Signed-Off-By: Jon Mason <jon@opengridcomputing.com>