Allow iser to be compiled under RHEL5.4
Add kernel of RHEL5.4 to the list of supported kernels (nackport patches
were already submitted by Or Gerlitz) and break the dependency between iSER and
iSCSI. However, the script code that builds iSCSI modules and upgrades
open-iscsi userspace RPM is still there in case we need it in the future.
Patches are to the install script and build scripts and were split into three parts.

Signed-off-by: Moni Shoua <monis@voltaire.com>