install.pl: Uninstall OFED RPMs coming with Linux Distribution.
install script includes the list of RPMs coming with different Linux
Distributions in order to provide the list of installed RPMs to user for
the manual uninstall as OFED may fail to uninstall these RPMs because of their
dependencies.
This patch makes this list to be used for all Distributions and not only for
SLES11.

Signed-off-by: Vladimir Sokolovsky <vlad@mellanox.co.il>